requirements.txt
Django==3.0
django-autoslug-iplweb==1.9.4
gunicorn==20.0.4
Pillow==6.2.1
Pillow-PIL==0.1.dev0
asgiref==3.2.3
pip==19.3.1
psycopg2==2.8.4
pytz==2019.3
setuptools==42.0.2
whitenoise==3.3.1
sqlparse==0.3.0